% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointly_estimate.R
\name{jointly_estimate}
\alias{jointly_estimate}
\title{Jointly estimate I0 and Rt}
\usage{
jointly_estimate(window, window_back, incid, si_distr, ...)
}
\arguments{
\item{window}{length of the window used for estimation. If you
have the full incidence time-series, only data in this window are
used for estimatiom.}

\item{window_back}{length of time for which incidence should be
estimated.}

\item{incid}{Incidence used to fit the model. Numeric vector of
length \code{window}}

\item{si_distr}{Discretised serial interval distribution}

\item{...}{Arguments passed to \code{rstan::fit} (e.g. iter, chains).}
}
\value{
an object of class stanfit
}
\description{
Joint estimation of initial incidence and reproduction number
}
\author{
Sangeeta Bhatia
}
